refactor: move generate_qrcode helper to top-level shared module
Both danfse and damdfe were importing draw_qr_code through
brazilfiscalreport.dacte.generate_qrcode, which forced importing the
DACTE module just to reach a shared helper. Moves the helper to
brazilfiscalreport/generate_qrcode.py and updates the three callers.
