+++ brazilfiscalreport/generate_qrcode.py
